Update and rename Go library
