+++ dp-hello-world-library-go/README.md
@@ -1,5 +1,5 @@
-# dp-hello-world-library
-Example library project generated by dp-cli
+# dp-hello-world-library-go
+Example go library project generated by dp-cli
 
 ### Getting started
 
@@ -15,7 +15,7 @@ See [CONTRIBUTING](CONTRIBUTING.md) for details.
 
 ### License
 
-Copyright © 2023, Office for National Statistics (https://www.ons.gov.uk)
+Copyright © 2025, Office for National Statistics (https://www.ons.gov.uk)
 
 Released under MIT license, see [LICENSE](LICENSE.md) for details.
