+++ dp-hello-world-library-go/ci/build.yml
@@ -9,7 +9,7 @@ image_resource:
     tag: latest
 
 inputs:
-  - name: dp-hello-world-library
+  - name: dp-hello-world-library-go
 
 outputs:
   - name: build
@@ -18,4 +18,4 @@ caches:
   - path: go/
 
 run:
-  path: dp-hello-world-library/ci/scripts/build.sh
+  path: dp-hello-world-library-go/ci/scripts/build.sh
