+++ dp-hello-world-library-go/ci/lint.yml
@@ -9,7 +9,7 @@ image_resource:
     tag: latest
 
 inputs:
-  - name: dp-hello-world-library
+  - name: dp-hello-world-library-go
 
 run:
-  path: dp-hello-world-library/ci/scripts/lint.sh
+  path: dp-hello-world-library-go/ci/scripts/lint.sh
