+++ dp-hello-world-library-go/go.mod
@@ -1,0 +1,11 @@
+module github.com/ONSdigital/dp-hello-world-library-go
+
+go 1.23
+
+require github.com/smartystreets/goconvey v1.8.1
+
+require (
+	github.com/gopherjs/gopherjs v1.17.2 // indirect
+	github.com/jtolds/gls v4.20.0+incompatible // indirect
+	github.com/smarty/assertions v1.15.0 // indirect
+)
