--- dp-hello-world-library/ci/scripts/lint.sh
@@ -1,5 +1,0 @@
-#!/bin/bash -eux
-
-pushd dp-hello-world-library
-  make lint
-popd
