--- dp-hello-world-library/go.mod
@@ -1,12 +1,0 @@
-module github.com/ONSdigital/dp-hello-world-library
-
-go 1.19
-
-require github.com/smartystreets/goconvey v1.7.2
-
-require (
-	github.com/gopherjs/gopherjs v0.0.0-20181017120253-0766667cb4d1 // indirect
-	github.com/jtolds/gls v4.20.0+incompatible // indirect
-	github.com/smartystreets/assertions v1.2.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
-)
